Add MySql db and Cookies

+++ src/controllers/CartControllers.kt
@@ -1,17 +1,43 @@
 package controllers
 
+import domain.cart.UserProductCart
 import infrastructure.NotFound
 import jakarta.servlet.annotation.WebServlet
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
 import org.thymeleaf.context.Context
 
+fun HttpServletRequest.getCartOrAnonymous(): MutableList<UserProductCart> {
+    val session = this.getSession(true)
+    var existingCart = session.getAttribute("userCart")
+
+    if (existingCart == null || existingCart !is List<*>) {
+        existingCart = mutableListOf<UserProductCart>()
+        this.setAttribute("userCart", existingCart)
+        return existingCart
+    }
+
+    return existingCart.filterIsInstance<UserProductCart>().toMutableList()
+}
+
+fun HttpServletRequest.setCart(cart: List<UserProductCart>) {
+    val session = this.getSession(true)
+    session.setAttribute("userCart", cart)
+}
+
 @WebServlet("/cart")
 class CartsServlet : BaseController() {
     override fun get(req: HttpServletRequest, resp: HttpServletResponse, context: Context): String {
-        val user = req.getUserOrAnonymous()
+        val user = req.getUser()
 
-        context.setVariable("cart", cartRepository.getUserCart(user.id))
+        if (user == null) {
+            val cart = req.getCartOrAnonymous()
+            val products = this.productRepository.getAllProducts()
+            val cartProducts = cart.map { it.productId }.map { products.find { p -> p.sku == it } }
+            context.setVariable("cart", cartProducts)
+        } else {
+            context.setVariable("cart", cartRepository.getUserCart(user.id))
+        }
 
         return "cart"
     }
@@ -23,33 +49,50 @@ class CartServlet : BaseController() {
     override fun get(req: HttpServletRequest, resp: HttpServletResponse, context: Context): String? {
         val slug = req.getRouteParameter()
 
-        if (slug == "add") {
-            return "products/add"
-        } else if (slug != null) {
+        return if (slug == "add") {
+            "products/add"
+        } else {
             context.setVariable("product", productRepository.getProductBySlug(slug))
 
-            return "products/product"
+            "products/product"
         }
-
-        return null
     }
 
     override fun post(req: HttpServletRequest, resp: HttpServletResponse) {
-        val user = req.getUserOrAnonymous()
+        val user = req.getUser()
 
-        val sku = req.getRouteParameter()?.toIntOrNull() ?: throw NotFound()
+        val sku = req.getRouteParameter().toIntOrNull() ?: throw NotFound()
 
-        cartRepository.addProductToCart(user.id, sku)
+        if (user == null) {
+            val cart = req.getCartOrAnonymous()
+
+            val existingQuantity = cart.find { it.productId == sku }?.quantity ?: 0
+            cart.removeIf { it.productId == sku }
+
+            cart.add(UserProductCart(sku, existingQuantity + 1))
+
+            req.setCart(cart)
+        } else {
+            cartRepository.addProductToCart(user.id, sku)
+        }
 
         resp.sendRedirect("/cart")
     }
 
     override fun delete(req: HttpServletRequest, resp: HttpServletResponse) {
-        val user = req.getUserOrAnonymous()
+        val user = req.getUser()
 
-        val sku = req.getRouteParameter()?.toIntOrNull() ?: throw NotFound()
+        val sku = req.getRouteParameter().toIntOrNull() ?: throw NotFound()
 
-        cartRepository.removeProductFromCart(user.id, sku)
+        if (user == null) {
+            val cart = req.getCartOrAnonymous()
+
+            cart.removeIf { it.productId == sku }
+
+            req.setCart(cart)
+        } else {
+            cartRepository.removeProductFromCart(user.id, sku)
+        }
 
         resp.setHeader("HX-Refresh", "true")
     }
